# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CoachMessageRole = typeof CoachMessageRole[keyof typeof CoachMessageRole];


export const CoachMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface CoachMessage {
  role: CoachMessageRole;
  /** @minLength 1 */
  content: string;
}

export interface CoachChatInput {
  /** @minItems 1 */
  messages: CoachMessage[];
}

export interface CoachChatReply {
  reply: string;
}

export interface ErrorResponse {
  error: string;
}