# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: coach
    description: AI Coach operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /coach/chat:
    post:
      operationId: sendCoachMessage
      tags: [coach]
      summary: Ask the AI digital safety coach
      description: Sends the conversation so far to the AI coach and returns its reply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachChatInput"
      responses:
        "200":
          description: Coach reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachChatReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Coach failed to respond
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CoachMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          minLength: 1
      required:
        - role
        - content
    CoachChatInput:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CoachMessage"
      required:
        - messages
    CoachChatReply:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error