# lib/api-zod/src/generated/types/coachChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachMessage } from './coachMessage';

export interface CoachChatInput {
  /** @minItems 1 */
  messages: CoachMessage[];
}